Rework ViewListener to allow working with generators.
In this first step, the ViewListener does not actually convert arrays to
rendered templates anymore but it delegates to a new interface
called `ControllerResultConverter`, where the templating code is moved
to an implementation called `ArrayToTemplateResponseConverter`.

The `RedirectListener` code is then added to another
`RedirectConverter`.

In the next step we need to add tests for both converters and then also
add other converters such as `FlashMessageConverter`,
`SwiftMessageConverter`, caching related converters/responses and such.

The idea is to return multiple results from controllers that are
combined together:

    public function listAction()
    {
        yield ['articles' => $this->repository->findAll()];
        yield new ExpiresCache(new DateTime('+600 seconds'));
    }

    public function editAction()
    {
        yield new RedirectRoute('list');
        yield new FlashMessage('Edited Article');
    }